+++ src/main/groovy/com/jakewharton/sdkmanager/SdkManagerExtension.groovy
@@ -3,4 +3,5 @@ package com.jakewharton.sdkmanager;
 class SdkManagerExtension {
   String emulatorVersion
   String emulatorArchitecture
+  String minSdkToolsVersion
 }