+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Version 1.5.0 *(In Development)*
 --------------------------------
 
  * New: Support for the 1.5.x Android plugin.
+ * New: SDK Tools will be updated if version is older than defined with sdkManager.minSdkToolsVersion
+   in build.gradle:
+      sdkManager {
+         minSdkToolsVersion '24.3.4'
+      }
  * New: Download r24.2 Android SDK.